Short values are now stored as int values.

+++ src/api/web/gw2/mapping/v2/characters/inventory/JsonInventoryUtils.java
@@ -62,7 +62,7 @@ public enum JsonInventoryUtils {
     public Inventory jsonObjectToInventory(final JsonObject jsonObject) {
         final DefaultInventory result = new DefaultInventory();
         result.id = jsonObject.getInt("id"); // NOI18N.
-        result.count = (short) jsonObject.getInt("count"); // NOI18N.
+        result.count = jsonObject.getInt("count"); // NOI18N.
         return result;
     }
 }
