+++ test/api/web/gw2/mapping/v2/colors/JsonpColorRGBTest.java
@@ -39,7 +39,7 @@ public class JsonpColorRGBTest {
         instance = new JsonpColorRGB();
         final Field rgbField = instance.getClass().getDeclaredField("rgb"); // NI18N.
         rgbField.setAccessible(true);
-        final short[] rgb = (short[]) rgbField.get(instance);
+        final int[] rgb = (int[]) rgbField.get(instance);
         rgb[0] = 88;
         rgb[1] = 97;
         rgb[2] = 200;
@@ -58,7 +58,7 @@ public class JsonpColorRGBTest {
         System.out.println("getRed");
         final Field rgbField = instance.getClass().getDeclaredField("rgb"); // NI18N.
         rgbField.setAccessible(true);
-        final short[] rgb = (short[]) rgbField.get(instance);
+        final int[] rgb = (int[]) rgbField.get(instance);
         final int expResult = rgb[0];
         final int result = instance.getRed();
         assertEquals(expResult, result);
@@ -72,7 +72,7 @@ public class JsonpColorRGBTest {
         System.out.println("getGreen");
         final Field rgbField = instance.getClass().getDeclaredField("rgb"); // NI18N.
         rgbField.setAccessible(true);
-        final short[] rgb = (short[]) rgbField.get(instance);
+        final int[] rgb = (int[]) rgbField.get(instance);
         final int expResult = rgb[1];
         final int result = instance.getGreen();
         assertEquals(expResult, result);
@@ -86,7 +86,7 @@ public class JsonpColorRGBTest {
         System.out.println("getBlue");
         final Field rgbField = instance.getClass().getDeclaredField("rgb"); // NI18N.
         rgbField.setAccessible(true);
-        final short[] rgb = (short[]) rgbField.get(instance);
+        final int[] rgb = (int[]) rgbField.get(instance);
         final int expResult = rgb[2];
         final int result = instance.getBlue();
         assertEquals(expResult, result);
